requirements.txt
django==4.2.6
django-ninja==0.22.2
prefect==2.14.1
prefect-shell==0.2.0
dj-database-url
psycopg2
cython<3.0
pydantic
nextflowpy==0.8.1
